+++ NEWS.md
@@ -4,6 +4,7 @@
 * `tm_g_km` updated to allow plot of failure probability on y-axis.
 * New module `tm_g_patient_profile` was added to profile patients based on predefined categories.
 * Exposed hard-coded variables using `data_extract_input`.
+* New module `tm_g_ipp` was added for individual patient plots.
 
 # teal.modules.clinical 0.8.6